Se mueve listado de favoritos a un componente.
